requirements.txt
python-dotenv==1.0.0
requests==2.31.0
geopy==2.4.1
openai==0.27.0
python-dateutil==2.8.2
flask==3.0.2
flask-cors==4.0.0
